refactor folder interceptor in tasks
